+++ end_to_end_testing/setup_helm2.sh
@@ -1,7 +1,7 @@
 source "$( cd -P "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"/setup_common.sh
 
 echo "Installing Helm"
-curl -sL https://storage.googleapis.com/kubernetes-helm/helm-v2.14.2-linux-amd64.tar.gz | tar xzv linux-amd64/helm
+curl -sL https://storage.googleapis.com/kubernetes-helm/helm-v2.16.1-linux-amd64.tar.gz | tar xzv linux-amd64/helm
 sudo mv linux-amd64/helm /usr/local/bin/helm
 rm -rf linux-amd64
 helm version --client